+++ extensions/brew/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Improvements] - 2024-06-24
 
+- Search now includes cask names
+
+## [Improvements] - 2024-06-24
+
 - Move formula and cask info to the metadata panel
 - Dependencies are displayed green if installed